requirements.txt
streamlit
numpy
pillow
imageio
